Move all commands to OverkizServer
