+++ pyoverkiz/servers/nexity.py
@@ -17,7 +17,7 @@ NEXITY_COGNITO_REGION = "eu-west-1"
 
 
 class NexityServer(OverkizServer):
-    async def login(self, username: str, password: str) -> bool:
+    async def _login(self, username: str, password: str) -> bool:
         """
         Authenticate and create an API session allowing access to the other operations.
         Caller must provide one of [userId+userPassword, userId+ssoToken, accessToken, jwt]
